src/test/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["sobottju <julian.benedikt.sobott@hs-furtwangen.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"
pydantic = "^1.10.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
